requirements.txt
openai>=1.0.0
Pillow>=10.0.0
